backend/requirements.txt
# GridAgent Backend Orchestrator Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# E2B Sandbox SDK
e2b>=0.17.0

# Supabase client (for future persistence)
supabase>=2.3.0

# JWT verification
PyJWT>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for potential external calls)
httpx>=0.26.0

# Pydantic for data validation (included with FastAPI but pinned for compatibility)
pydantic>=2.5.0
